requirements.txt
selenium==3.141.0
pytest==4.6.2
allure-pytest==2.6.5
pytest-parallel==0.0.9
junitxml==0.7